[FLINK-4315] Hadoop dependencies were moved from flink-java and flink-scala
