+++ flink-tests/src/test/scala/org/apache/flink/api/scala/hadoop/mapreduce/WordCountMapreduceITCase.scala
@@ -19,6 +19,7 @@
 package org.apache.flink.api.scala.hadoop.mapreduce
 
 import org.apache.flink.api.scala._
+import org.apache.flink.api.scala.hadoop.FlinkHadoopEnvironment
 import org.apache.flink.test.testdata.WordCountData
 import org.apache.flink.test.util.{TestBaseUtils, JavaProgramTestBase}
 import org.apache.hadoop.fs.Path
@@ -45,7 +46,8 @@ class WordCountMapreduceITCase extends JavaProgramTestBase {
     val env = ExecutionEnvironment.getExecutionEnvironment
 
     val input =
-      env.readHadoopFile(new TextInputFormat, classOf[LongWritable], classOf[Text], textPath)
+      FlinkHadoopEnvironment.getHadoopEnvironment(env).
+        readHadoopFile(new TextInputFormat, classOf[LongWritable], classOf[Text], textPath)
 
     val text = input map { _._2.toString }
     val counts = text.flatMap { _.toLowerCase.split("\\W+") filter { _.nonEmpty } }
